+++ static/index.html
@@ -10,15 +10,13 @@
   <body>
     <h1>List of Problems </h2>
       <h3>  OPTION      PROBLEM TYPE
-  <ol>
-  <li>Translate Random Virtual Address to a Physical Address </li>
-  <li>Translate Random Virtual Address Returned by Malloc() to a Physical Address
-  <li>Calculate Specified Portion through the Stack as a Virtual Address
-  <li>Generate fresh segmented memory model
-  <li>Generate Random Problem
-  </ol>
+  <li>0     Translate Random Virtual Address to a Physical Address
+  <li>1     Translate Random Virtual Address Returned by Malloc() to a Physical Address
+  <li>2     Calculate Specified Portion through the Stack as a Virtual Address
+  <li>8     Generate fresh segmented memory model
+  <li>9     Generate Random Problem
     <form action="/first" method="get" accept-charset="utf-8">
-      <input id="formBox" type="text" name="term" placeholder="Type Here"/>
+      <input id="formBox" type="text" name="question_format" placeholder="Type Here"/>
       <input type="submit" value="Next">
     </form>
