+++ static/stack_format.html
@@ -11,18 +11,9 @@
     <h1> Choose format of desired question </h1>
 
             <h3>OPTION          PROBLEM TYPE
-              <ol>
-            <li>--va in hex to pa in hex
-            <li>--va in hex to pa in binary
-            <li>--va in hex to pa in decimal
-            <li>--va in binary to pa in hex
-            <li>--va in binary to pa in binary
-            <li>--va in binary to pa in decimal
-            <li>--va in decimal to pa in hex
-            <li>--va in decimal to pa in binary
-            <li>--va in decimal to pa in decimal
-            <li>Random option
-            </ol>
+            <li>0 --va in hex to pa in hex
+            <li>1 --va in hex to pa in binary
+            <li>2 --va in hex to pa in decimal
     <div>
       <br>
     <form action="/search" method="post" accept-charset="utf-8" action="result.html">
