+++ static/va_to_pa_format.html
@@ -1,0 +1,32 @@
+<!DOCTYPE html>
+<html lang="en">
+
+  <head>
+    <meta charset="utf-8">
+    <title>Index</title>
+    <link rel="stylesheet" href="/static/index.css">
+  </head>
+
+  <body>
+    <h1> Choose format of desired question </h1>
+
+            <h3>OPTION          PROBLEM TYPE
+            <li>0 --va in hex to pa in hex
+            <li>1 --va in hex to pa in binary
+            <li>2 --va in hex to pa in decimal
+            <li>3 --va in binary to pa in hex
+            <li>4 --va in binary to pa in binary
+            <li>5 --va in binary to pa in decimal
+            <li>6 --va in decimal to pa in hex
+            <li>7 --va in decimal to pa in binary
+            <li>8 --va in decimal to pa in decimal
+            <li>9 Random option
+    <div>
+      <br>
+    <form action="/search" method="post" accept-charset="utf-8" action="result.html">
+      <input id="formBox" type="text" name="term" placeholder="Type Here"/>
+      <input type="submit" value="Next">
+    </form>
+  </div>
+  </body>
+</html>
